extract cell types
