+++ frontend/fingertips-frontend/components/organisms/Heatmap/components/cell/dataCell/dataCell.test.tsx
@@ -1,43 +1,6 @@
-import { render } from '@testing-library/react';
-import { HeatmapCell } from '.';
-import { CellType } from '../../heatmapUtil';
 import { GovukColours } from '@/lib/styleHelpers/colours';
-
-describe('cell types', () => {
-  it.each([
-    [
-      'indicator title',
-      CellType.IndicatorTitle,
-      'heatmap-cell-indicator-title',
-    ],
-    [
-      'indicator information period',
-      CellType.IndicatorInformationPeriod,
-      'heatmap-cell-indicator-info-period',
-    ],
-    [
-      'indicator information value unit',
-      CellType.IndicatorInformationValueUnit,
-      'heatmap-cell-indicator-info-value-unit',
-    ],
-    ['data', CellType.Data, 'heatmap-cell-data'],
-  ])(
-    'should render appropriate type for %s cell',
-    (_, cellType, expectedTestId) => {
-      const screen = render(
-        <table>
-          <tbody>
-            <tr>
-              <HeatmapCell cellType={cellType} content={''} />
-            </tr>
-          </tbody>
-        </table>
-      );
-
-      expect(screen.getByTestId(expectedTestId)).toBeInTheDocument();
-    }
-  );
-});
+import { render } from '@testing-library/react';
+import { DataCell } from './dataCell';
 
 describe('data cell', () => {
   it('should render white background with black text by default', () => {
@@ -45,7 +8,7 @@ describe('data cell', () => {
       <table>
         <tbody>
           <tr>
-            <HeatmapCell cellType={CellType.Data} content={''} />
+            <DataCell content={''} />
           </tr>
         </tbody>
       </table>
@@ -64,11 +27,7 @@ describe('data cell', () => {
       <table>
         <tbody>
           <tr>
-            <HeatmapCell
-              cellType={CellType.Data}
-              content={''}
-              backgroundColour={expectedColour}
-            />
+            <DataCell content={''} backgroundColour={expectedColour} />
           </tr>
         </tbody>
       </table>
@@ -86,8 +45,7 @@ describe('data cell', () => {
       <table>
         <tbody>
           <tr>
-            <HeatmapCell
-              cellType={CellType.Data}
+            <DataCell
               content={''}
               backgroundColour={expectedBackgroundColour}
             />
