+++ frontend/fingertips-frontend/components/organisms/Heatmap/components/cell/indicatorCell/indicatorInfoValueUnitCell.tsx
@@ -1,9 +1,8 @@
 import { FC } from 'react';
 import {
-  StyledCellText,
   StyledDivIndicatorInformationValueUnitCellContent,
-} from './heatmapCell.styles';
-
+  TextCell,
+} from './indicatorCell.styles';
 interface IndicatorInfoValueUnitCellProps {
   content: string;
 }
@@ -12,10 +11,10 @@ export const IndicatorInfoValueUnitCell: FC<
   IndicatorInfoValueUnitCellProps
 > = ({ content }: IndicatorInfoValueUnitCellProps) => {
   return (
-    <StyledCellText data-testid="heatmap-cell-indicator-info-value-unit">
+    <TextCell data-testid="heatmap-cell-indicator-info-value-unit">
       <StyledDivIndicatorInformationValueUnitCellContent>
         {content}
       </StyledDivIndicatorInformationValueUnitCellContent>
-    </StyledCellText>
+    </TextCell>
   );
 };
