+++ frontend/fingertips-frontend/components/organisms/Heatmap/components/cell/indicatorCell/indicatorTitleCell.tsx
@@ -1,0 +1,19 @@
+import { FC } from 'react';
+import {
+  IndicatorTitleCellContent,
+  IndicatorCell,
+} from './indicatorCell.styles';
+
+interface IndicatorTitleCellProps {
+  content: string;
+}
+
+export const IndicatorTitleCell: FC<IndicatorTitleCellProps> = ({
+  content,
+}: IndicatorTitleCellProps) => {
+  return (
+    <IndicatorCell data-testid="heatmap-cell-indicator-title">
+      <IndicatorTitleCellContent>{content}</IndicatorTitleCellContent>
+    </IndicatorCell>
+  );
+};
